Remove funsor.six module
