+++ funsor/interpreter.py
@@ -4,6 +4,7 @@ import os
 import re
 import types
 from collections import OrderedDict
+from functools import singledispatch
 
 import numpy
 import torch
@@ -12,7 +13,6 @@ from contextlib2 import contextmanager
 from funsor.domains import Domain
 from funsor.ops import Op
 from funsor.registry import KeyedRegistry
-from funsor.six import singledispatch
 
 _ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 _DEBUG = int(os.environ.get("FUNSOR_DEBUG", 0))
